+++ packages/mg-squarespace-xml/lib/process.js
@@ -1,6 +1,20 @@
 const $ = require('cheerio');
+const {slugify} = require('@tryghost/string');
 
-const processContent = (html) => {
+module.exports.processUser = ($sqUser) => {
+    const authorSlug = slugify($($sqUser).children('wp\\:author_login').text());
+
+    return {
+        url: authorSlug,
+        data: {
+            slug: authorSlug,
+            name: $($sqUser).children('wp\\:author_display_name').text(),
+            email: $($sqUser).children('wp\\:author_email').text()
+        }
+    };
+};
+
+module.exports.processContent = (html) => {
     if (!html) {
         return '';
     }
@@ -69,20 +83,171 @@ const processContent = (html) => {
     return html;
 };
 
-const processPost = (post, siteUrl) => {
-    post.data.html = processContent(post.data.html, siteUrl);
+// The feature images is not "connected" to the post, other than it's located
+// in the sibling `<item>` node.
+module.exports.processFeatureImage = ($sqPost) => {
+    const $nextItem = $($sqPost).next().children('wp\\:attachment_url');
 
-    return post;
+    if ($nextItem.length >= 1) {
+        return $nextItem.text();
+    }
+
+    return;
 };
 
-module.exports = async (input, ctx) => {
-    let {options} = ctx;
-    let {url: siteUrl} = options;
-    const output = {};
+module.exports.processTags = ($sqCategories, fetchTags) => {
+    const categories = [];
+    const tags = [];
 
-    if (input.posts && input.posts.length > 0) {
-        output.posts = input.posts.map(post => processPost(post, siteUrl));
-        output.users = input.users;
+    $sqCategories.each((i, taxonomy) => {
+        if (fetchTags && $(taxonomy).attr('domain') === 'post_tag') {
+            tags.push({
+                url: `/tag/${$(taxonomy).attr('nicename')}`,
+                data: {
+                    slug: $(taxonomy).attr('nicename'),
+                    name: $(taxonomy).text()
+                }
+            });
+        } else if ($(taxonomy).attr('domain') === 'category') {
+            categories.push({
+                url: `/tag/${$(taxonomy).attr('nicename')}`,
+                data: {
+                    slug: $(taxonomy).attr('nicename'),
+                    name: $(taxonomy).text()
+                }
+            });
+        }
+    });
+
+    return categories.concat(tags);
+};
+
+module.exports.processPost = ($sqPost, users, {addTag, tags: fetchTags, siteUrl}) => {
+    const postType = $($sqPost).children('wp\\:post_type').text();
+
+    // only grab posts and pages
+    if (postType !== 'attachment') {
+        const featureImage = this.processFeatureImage($sqPost);
+        const authorSlug = slugify($($sqPost).children('dc\\:creator').text());
+        let postSlug = $($sqPost).children('link').text();
+
+        if (!postSlug || postSlug.indexOf('null') >= 0) {
+            // drafts can have a post slug/link of `/null`
+            postSlug = 'untitled';
+        }
+
+        const post = {
+            url: `${siteUrl}${$($sqPost).children('link').text()}`,
+            data: {
+                slug: $($sqPost).children('wp\\:post_name').text().replace(/(\.html)/i, ''),
+                title: $($sqPost).children('title').text(),
+                status: $($sqPost).children('wp\\:status').text() === 'publish' ? 'published' : 'draft',
+                // TODO: properly format the date, as `pubDate` is not accepted by our importer
+                published_at: $($sqPost).children('wp\\:post_date_gmt').text() || $($sqPost).children('pubDate').text(),
+                // TODO: clean the custom excerpt to not include any HTML
+                custom_excerpt: $($sqPost).children('excerpt\\:encoded').text(),
+                feature_image: featureImage,
+                tags: [],
+                type: postType,
+                author: users ? users.find(user => user.data.slug === authorSlug) : null
+            }
+        };
+
+        post.data.html = this.processContent($($sqPost).children('content\\:encoded').text());
+
+        if ($($sqPost).children('category').length >= 1) {
+            post.data.tags = this.processTags($($sqPost).children('category'), fetchTags);
+        }
+
+        post.data.tags.push({
+            url: 'migrator-added-tag', data: {name: '#sqs'}
+        });
+
+        if (addTag) {
+            post.data.tags.push({
+                url: 'migrator-added-tag-2', data: {slug: addTag, name: addTag}
+            });
+        }
+
+        if (!post.data.author) {
+            if ($($sqPost).children('dc\\:creator').length >= 1) {
+                post.data.author = {
+                    url: authorSlug,
+                    data: {
+                        slug: authorSlug
+                    }
+                };
+            } else {
+                post.data.author = {
+                    url: 'migrator-added-author',
+                    data: {
+                        slug: 'migrator-added-author'
+                    }
+                };
+            }
+        }
+
+        return post;
+    }
+
+    return;
+};
+
+module.exports.processPosts = ($xml, users, options) => {
+    const postsOutput = [];
+
+    $xml('item').each((i, sqPost) => {
+        postsOutput.push(this.processPost(sqPost, users, options));
+    });
+
+    // don't return empty post objects
+    return postsOutput.filter(post => post);
+};
+
+module.exports.processUsers = ($xml) => {
+    const usersOutput = [];
+
+    $xml('wp\\:author').each((i, sqUser) => {
+        usersOutput.push(this.processUser(sqUser));
+    });
+
+    return usersOutput;
+};
+
+module.exports.all = async (input, {options}) => {
+    const {drafts, pages} = options;
+    const output = {
+        posts: [],
+        users: []
+    };
+
+    if (input.length < 1) {
+        return new Error('Input file is empty');
+    }
+
+    const $xml = $.load(input, {
+        decodeEntities: false,
+        xmlMode: true,
+        lowerCaseTags: true // needed to find `pubDate` tags
+    });
+
+    // grab the URL of the site we're importing
+    options.siteUrl = $xml('channel > link').text();
+
+    // process users first, as we're using this information
+    // to populate the author data for posts
+    output.users = this.processUsers($xml);
+
+    output.posts = this.processPosts($xml, output.users, options);
+
+    if (!drafts) {
+        // remove draft posts
+        output.posts = output.posts.filter(post => post.data.status !== 'draft');
+    }
+
+    if (!pages) {
+        // remove pages. absolute not supported by default and not tested!!
+        output.posts = output.posts.filter(post => post.data.type !== 'page');
     }
 
     return output;
